+++ data/src/main/java/com/example/data/remote/api/TMDBApi.kt
@@ -19,7 +19,7 @@ interface TMDBApi {
     suspend fun fetchUpcomingMovies(
             @Query("api_key") apiKey: String,
             @Query("language") language: String,
-            @Query("page") page: Long,
+            @Query("page") page: Int,
             @Query("region") region: String
     ): UpcomingMoviesResponseModel
